v_video: Drop V_DrawMouseSpeedBox
This contains a call to fabs() from math.h, which libpayload does not
implement. V_DrawMouseSpeedBox was used to show the current mouse speed,
and mouse support isn't currently implemented in coreDOOM. This was
never in the original Doom source code release anyway, it was added in
one of the derivatives at some point.

Signed-off-by: Nicholas Chin <nic.c3.14@gmail.com>